% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_DNA.R
\name{form_DNA}
\alias{form_DNA}
\title{Title form_DNA}
\usage{
form_DNA(size)
}
\arguments{
\item{size}{integer value that represent the length of the DNA sequence to be randomly generated.}
}
\value{

}
\description{
Title form_DNA
}
