% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F3_get_codons.R
\name{get_codons}
\alias{get_codons}
\title{Title get_codons}
\usage{
get_codons(mRNA_seq, start = 1)
}
\arguments{
\item{start}{it's an integer that represents the starting point of first the reading frame .}

\item{DNA_seq}{it's a string  that represents a DNA sequence to be transcribed  with the transcribe function  .}
}
\value{
codons which is vector of three chars presenting each codon from the DNA_seq starting from the start .
}
\description{
Title get_codons
}
\examples{
get_codons("AUGC", start =  1)

}
