% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nucleotide_destribution.R
\name{nucleotide_destribution}
\alias{nucleotide_destribution}
\title{nucleotide destribution}
\usage{
nucleotide_destribution(nucleotide_sequences)
}
\arguments{
\item{nucleotide_sequences}{Matrix with vectors of nucleotides sequences}
}
\value{
A plot with the destribution of nucleotides in each sequence
}
\description{
nucleotide destribution
}
