% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F5_plot_base_occurrence.R
\name{plot_base_occurrence}
\alias{plot_base_occurrence}
\title{Plots the occurrence of bases in DNA/RNA sequence}
\usage{
plot_base_occurrence(sequence)
}
\arguments{
\item{sequence}{a string with characters (bases) to be counted}
}
\value{
ggplot object displaying the total count per character (base)
}
\description{
Plots the occurrence of bases in DNA/RNA sequence
}
