% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rna_codons}
\alias{rna_codons}
\title{RNA Codons}
\format{

}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
rna_codons
}
\description{
List with RNA codons encoding Amino Acids
}
\keyword{datasets}
