% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnd_dna_sample.R
\name{rnd_dna_sample}
\alias{rnd_dna_sample}
\title{Random DNA sample}
\usage{
rnd_dna_sample(dna_size)
}
\arguments{
\item{dna_size}{Given integer for DNA sample size}
}
\value{
Returns a random DNA sample in string format
}
\description{
Random DNA sample
}
\examples{
rnd_dna_sample(10)
rnd_dna_sample(15)
rnd_dna_sample(20)
}
