% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_all.R
\name{select_ribo}
\alias{select_ribo}
\title{select_ribo}
\usage{
select_ribo(codons)
}
\arguments{
\item{codons}{Selects the ribos codons, with a "" between them}
}
\value{
returns ribosomes
}
\description{
select_ribo
}
\examples{
ribo <- c("AUG", "GCU", "UAA", "CGG")
codons <- c(1, 2, 4)
select_ribo(codons)
returns "AUGGCUCGG"
}
