% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_reading_frame.R
\name{set_reading_frame}
\alias{set_reading_frame}
\title{set_reading_frame}
\usage{
set_reading_frame(RNA, start = 1)
}
\arguments{
\item{RNA}{Give the RNA in a form of a string containing the nucleutide sequence}

\item{start}{Give the location of the +1 nucleutide in the string.}
}
\value{
An ordered list of codons starting from the defined nucleutide.
}
\description{
set_reading_frame
}
\examples{

mRNA <- 'CAUGUUA'
codons <- set_reading_frame(RNA = mRNA, start = 2)

}
