% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{transcribe}
\alias{transcribe}
\title{Title Transcription}
\usage{
transcribe(DNA_seq)
}
\arguments{
\item{DNA_seq}{}
}
\value{
mRNA
}
\description{
Title Transcription
}
\examples{
transcribe ("ATCG")
}
