% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Title Translation}
\usage{
translate(codons)
}
\arguments{
\item{codons:}{set of mRNA (3 mRNA turn to a amino_acid)}
}
\value{
translated protein from mRNA codon
}
\description{
Title Translation
}
