% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1_form_DNA.R
\name{form_dna}
\alias{form_dna}
\title{Title form_DNA}
\usage{
form_dna(size = 1)
}
\arguments{
\item{size}{integer value that represent the length of the DNA sequence to be randomly generated}
}
\value{
A sequence of nucleotide base as random_dna2
}
\description{
Title form_DNA
}
\examples{
form_dna(size = 15)

}
