% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codon.R
\name{get_codons}
\alias{get_codons}
\title{Title get_codons}
\usage{
get_codons(DNA_seq, start = 1)
}
\arguments{
\item{DNA_seq}{it's a string  that represents a DNA sequence to be transcribed  with the transcribe function  .}

\item{start}{it's an integer that represents the starting point of first the reading frame .}
}
\value{
codons which is vector of three chars presenting each codon from the DNA_seq starting from the start .
}
\description{
Title get_codons
}
\examples{
get_codons("ATGC", start =  1)

}
